pseudo-Ballesteros-Weinstein
BW is the standard for rhodopsin family members, but application to
V1Rs is problematic: the corresponding "50-position" residues are not
conserved, and in most cases they are not even the most conserved
residue of their helix (among V1Rs).

This takes the approach of using an explicit choice of
"50-position" residue on each TM helix. In conjunction with
structure-based  alignment, this allows one to build correspondences
to the *rhodopsin* labeling scheme regardless of conservation among
V1Rs.

A suggested notation is X123^r5.50, where X is the AA,
123 is its position in a particular V1R, ^ indicates exponent,
and `r5.50` mean residue that best matches (in the 3d structure)
the rhodopsin 5.50 residue.