Download latest AlphaFold (#7)
There appear to be multiple versions of the same structures;
currently they appear to be on 'v4'. While there doesn't seem
to be a way to query the database as a whole for the latest
version, we can start at 4 and work our way backwards.

Also puts the download under a progress meter.