Add Needleman-Wunsch alignment
This is a special case in which no gaps are allowed in one of the
"sequences." The intent is to use this for aligning the TM helices of
GPCRs: the "reference sequence" is just the 14 residues at the start and
end of the reference structure's TM helices, and the "query sequence" is
the full sequence of another 7TM.

The core algorithm is supplied with a pairwise cost matrix. Convenience
methods allow you to compute the cost from sequence. Currently, cost can
be computed either from distance alone or distance + orientation of the
residues.