Generalize Needleman-Wunsch alignment (#16)
The initial application of NW alignment was to find TM anchors. In that
case, one is aligning a "probe" structure to 14 specific "reference
residues," and thus the problem is quite asymmetric. There are
applications where we would rather do "symmetric" alignment. This is
basically a complete rewrite of the NW alignment code to allow for this.

With the centroid/sidechain distance-based penalty, recommended gapcosts
are: 5Å for extend, 20Å for open. These approximately maximize the
agreement with the MSA pairings.