Add `forcecomponents`, generalize features
This breaking change adds a 3rd argument to the functions that compute
`σ` and `ϕ`, requiring the feature name as an argument. This allows one
to tune the decay-length or amplitude based on specific atom, residue,
and/or field-type.

This also establishes defaults for different fields that attempt to be a
bit more physically realistic. These have not, however, been
investigated in great detail.

Finally, `forcecomponents` can be used to compute the net force on a
residue, broken down by field. This is useful for tuning the amplitude,
using the condition that a realistic structure is in equilibrium.