Switch to BioStructures for PDB/structure rep
BioStructures is poised to become the standard package for protein
structure I/O and representation. While MIToS is retained for the MSA
functionality and a few very useful data tables (vanderwaals radii and
atom-level amino acid traits), otherwise this is a complete switch. It's
somewhat surprising how few changes were needed to the tests (some of
which were essentially cleanups that could have happened earlier).