Switch to BioStructures for PDB/structure rep (#26)
BioStructures is poised to become the standard package for protein
structure I/O and representation. MIToS is retained for the MSA code
and a few very useful data tables (vanderwaals radii and atom-level
amino acid traits), but otherwise this is a complete switch. It's
somewhat surprising how few changes were needed to the tests (some of
which were essentially cleanups that could have happened earlier).

This also adds a small amount of new functionality, and enhances
docstrings, comments, etc.