Improve support for multiple MSA formats (#44)
This enhances support for BioStockholm and adds support for aligned FASTA as MSA
formats. It also adds a few enhancements for structures encoded in multiple
formats (e.g., PDB or CIF) and supports flexibility in the return format from
querying the EBI Proteins API (e.g., to download files in FASTA format).