Add atom-in-cyclinder features (#45)
This adds a far simpler criterion than the "inside hull" for choosing
atoms and/or residues for inclusion in the feature GMM. This can be used
to cross-check a result to ensure it's not dependent on subtle
variations in angle.

This also tweaks `align_to_membrane` to pair more naturally with this
atom-selection method, and improves its docstring.